Добавлен метод для добавления диапазона Добавлен импорт данных их excel Исправлена ошибка с типом данных

+++ Scripts/1_create_table_weather.sql
@@ -7,7 +7,7 @@ CREATE TABLE Weather (
 	dew_point DECIMAL(4, 1) NOT NULL,	-- Точка росы (в цельсиях)
 	pressure SMALLINT NOT NULL,			-- Атмосферное давление (в мм рт.ст.)
 	wind_direction VARCHAR(10),			-- Направление ветра (например, З,ЮЗ,штиль)
-	wind_speed DECIMAL(4, 1),			-- Скорость ветра (в м/с)
+	wind_speed SMALLINT,			-- Скорость ветра (в м/с)
 	cloudiness SMALLINT,				-- Облачность (в %)
 	cloud_base_height INT,				-- Нижняя граница облачности (в м)
 	horizontal_visibility DECIMAL(4, 1),-- Горизонтальная видимость (в км)
